pyproject.toml
[project]
name = "cloud-ai-security-skills"
version = "0.8.1"
description = "Security skills for cloud and AI with repo-native and OCSF modes."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }

[tool.uv]
package = false

[dependency-groups]
dev = [
  "bandit[toml]>=1.9.4,<2",
  "defusedxml>=0.7.1,<1",
  "moto>=5,<6",
  "mypy>=1.11,<2",
  "pip-audit>=2.7,<3",
  "pre-commit>=3.8,<5",
  "pytest>=8.3,<10",
  "pytest-cov>=5,<8",
  "ruff>=0.6.9,<1",
]
mcp = [
  "pyyaml>=6,<7",
]
aws = [
  "boto3>=1.35,<2",
]
gcp = [
  "google-cloud-compute>=1.47.0,<2",
  "google-cloud-iam>=2.16,<3",
  "google-cloud-resource-manager>=1.17.0,<2",
  "google-cloud-storage>=2.18,<4",
]
azure = [
  "azure-identity>=1.17,<2",
  "azure-mgmt-network>=28,<31",
  "azure-mgmt-resource>=25,<26",
  "azure-mgmt-storage>=21,<25",
]
iam_departures = [
  "clickhouse-connect>=0.7,<1",
  "databricks-sql-connector>=4.2.5,<5",
  "google-api-python-client>=2,<3",
  "httpx>=0.27,<1",
  "snowflake-connector-python>=4.4.0,<5",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.bandit]
exclude_dirs = ["tests", ".venv", ".git", "__pycache__"]
skips = ["B101"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib -q"
testpaths = ["skills"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
# Skills use a flat src/<entrypoint>.py layout (e.g. src/detect.py, src/ingest.py)
# repeated across dozens of skill directories. Running `mypy skills/` discovers
# many same-named modules and aborts with "Duplicate module named 'detect'".
# The canonical entrypoint is `scripts/run_mypy.sh`, which invokes mypy once
# per skill src/ directory. `explicit_package_bases` + `namespace_packages`
# keep ad-hoc per-directory invocations working from the IDE.
python_version = "3.11"
explicit_package_bases = true
namespace_packages = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
exclude = "(^|/)(tests|infra)/"

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true
